+++ src/directLineStreaming.ts
@@ -8,9 +8,9 @@ import * as BFSE from 'botframework-streaming';
 import createDeferred from './createDeferred';
 import fetch from 'cross-fetch';
 
-import watchNetworkInformation from './streaming/watchNetworkInformation';
-import watchREST from './streaming/watchREST';
-import WebSocketClientWithWatchdog from './streaming/WebSocketClientWithWatchdog';
+import probeNetworkInformation from './streaming/probeNetworkInformation';
+import probeREST from './streaming/probeREST';
+import WebSocketClientWithProbe from './streaming/WebSocketClientWithProbe';
 
 import type { Deferred } from './createDeferred';
 import { Activity, ConnectionStatus, Conversation, IBotConnection, Media, Message } from './directLine';
@@ -30,20 +30,20 @@ interface DirectLineStreamingOptions {
   botAgent?: string;
 
   /**
-   * Sets the connection liveness probe for assisting detection of connection issues.
+   * Sets the network probe for assisting detection of connection issues.
    *
    * When the probe detects any connection issues, the bot connection will be closed and treated as an error.
    *
    * The probe is intended to assist `WebSocket`. Some implementations of `WebSocket` does not emit `error` event timely in case of connection issues.
    * This probe will help declaring connection outages sooner. For example, on iOS/iPadOS 15 and up, the newer "NSURLSession WebSocket" did not signal error on network change.
    *
-   * There are 3 ways to set the probe: `NetworkInformation` instance (using [Network Information API](https://developer.mozilla.org/en-US/docs/Web/API/Network_Information_API)), `object` (REST API watchdog), or `function`.
+   * There are 3 ways to set the probe: `NetworkInformation` instance (using [Network Information API](https://developer.mozilla.org/en-US/docs/Web/API/Network_Information_API)), `object` (REST API probe), or `function`.
    *
    * When the probe is an instance of `NetworkInformation`:
    *
-   * - When `change` event is received on the instance, the watchdog will treat it as a fault.
+   * - When `change` event is received on the instance, the probe will treat it as a fault.
    *
-   * When the probe is an object, it will watch the liveness of a long-polling service via HTTP GET:
+   * When the probe is an object, it will probe the liveness of a long-polling service via HTTP GET:
    *
    * - `url` is the URL of a HTTP GET long-polling service. The service must keep the connection for a period of time and returns HTTP 2xx when the time has passed.
    *   [RFC6202](https://www.rfc-editor.org/rfc/rfc6202) recommends the connection should be kept for 30 seconds.
@@ -53,13 +53,13 @@ interface DirectLineStreamingOptions {
    *
    * When the probe is a function:
    *
-   * - The function will be called when a connection is being established. Thus, a new liveness probe is needed.
+   * - The function will be called when a connection is being established. Thus, a new network probe is needed.
    * - The returned `AbortSignal` should be aborted as soon as the probe detects any connection issues.
    * - The function should create a new probe on every call and probe should not be reused.
    * - When a probe is no longer needed, the `AbortSignal` passed to the function will signal release of underlying resources.
    * - At any point of time, there should be no more than 1 probe active. The chat adapter will signal the release of probe before requesting for a new one.
    */
-  watchdog?:
+  networkProbe?:
     | NetworkInformation
     | {
         minimumInterval?: number;
@@ -148,32 +148,32 @@ export class DirectLineStreaming implements IBotConnection {
 
   private _botAgent = '';
 
-  #watchdog: ((init: { signal: AbortSignal }) => AbortSignal) | undefined;
+  #networkProbe: ((init: { signal: AbortSignal }) => AbortSignal) | undefined;
 
   constructor(options: DirectLineStreamingOptions) {
-    // Rectifies options.watchdog.
-    const watchdog = options?.watchdog;
+    // Rectifies options.probe.
+    const networkProbe = options?.networkProbe;
 
-    if (typeof watchdog === 'function') {
-      this.#watchdog = watchdog;
-    } else if (typeof watchdog === 'undefined') {
+    if (typeof networkProbe === 'function') {
+      this.#networkProbe = networkProbe;
+    } else if (typeof networkProbe === 'undefined') {
       // Intentionally left blank.
     } else if (
-      watchdog instanceof EventTarget ||
+      networkProbe instanceof EventTarget ||
       // We also accept `EventTargetLike`.
-      (typeof watchdog['addEventListener'] === 'function' && typeof watchdog['removeEventListener'] === 'function')
+      (typeof networkProbe['addEventListener'] === 'function' && typeof networkProbe['removeEventListener'] === 'function')
     ) {
-      this.#watchdog = ({ signal }: { signal: AbortSignal }) =>
-        watchNetworkInformation(watchdog as NetworkInformation, { signal });
+      this.#networkProbe = ({ signal }: { signal: AbortSignal }) =>
+        probeNetworkInformation(networkProbe as NetworkInformation, { signal });
     } else if (
-      (typeof watchdog.minimumInterval === 'number' || typeof watchdog.minimumInterval === 'undefined') &&
-      (typeof watchdog.url === 'string' || watchdog.url instanceof URL)
+      (typeof networkProbe.minimumInterval === 'number' || typeof networkProbe.minimumInterval === 'undefined') &&
+      (typeof networkProbe.url === 'string' || networkProbe.url instanceof URL)
     ) {
-      this.#watchdog = ({ signal }: { signal: AbortSignal }) =>
-        watchREST(watchdog.url, { minimumInterval: watchdog.minimumInterval, signal });
+      this.#networkProbe = ({ signal }: { signal: AbortSignal }) =>
+        probeREST(networkProbe.url, { minimumInterval: networkProbe.minimumInterval, signal });
     } else {
       throw new Error(
-        'botframework-directlinejs: "watchdog" option must be either a function returning an AbortSignal, an object, or undefined.'
+        'botframework-directlinejs: "networkProbe" option must be either a function returning an AbortSignal, an object, or undefined.'
       );
     }
 
@@ -399,11 +399,12 @@ export class DirectLineStreaming implements IBotConnection {
     // This promise will resolve when it is disconnected.
     return new Promise(async (resolve, reject) => {
       try {
-        const watchdog: AbortSignal | undefined =
-          typeof this.#watchdog === 'function' ? this.#watchdog({ signal: abortController.signal }) : undefined;
+        const probe: AbortSignal | undefined =
+          typeof this.#networkProbe === 'function' ? this.#networkProbe({ signal: abortController.signal }) : undefined;
 
-        this.streamConnection = new WebSocketClientWithWatchdog({
+        this.streamConnection = new WebSocketClientWithProbe({
           disconnectionHandler: resolve,
+          probe,
           requestHandler: {
             processRequest: streamingRequest => {
               // If `streamConnection` is still current, allow call to `processRequest()`, otherwise, ignore calls to `processRequest()`.
@@ -416,7 +417,6 @@ export class DirectLineStreaming implements IBotConnection {
             }
           },
           url: wsUrl,
-          watchdog
         });
 
         this.queueActivities = true;